requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
asgiref==3.6.0
    # via django
build==0.10.0
    # via pip-tools
click==8.1.3
    # via pip-tools
django==4.2
    # via -r requirements.in
django-timescaledb==0.2.13
    # via -r requirements.in
packaging==23.1
    # via build
pip-tools==6.13.0
    # via -r requirements.in
psycopg2==2.9.6
    # via -r requirements.in
pyproject-hooks==1.0.0
    # via build
sqlparse==0.4.3
    # via django
wheel==0.40.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
